[ui] add options to configure links to kibana&grafana
The UI server displays links to Kibana & Grafana on the left page of
every page. The target of these links are retrieved from the
configuration server, from the "links" document of the "system" database.

This commit makes it easier to configure the target of these links by
setting two environment variables, GRAFANA_UI_URL and KIBANA_UI_URL.

Signed-off-by: Xavier <xavier@reblaze.com>