Prevent empty tag fields for reason
The tag field in the reason field could be empty for challenge actions
that did not emanate from acl rules. This patch prevents the field from
being displayed if it is empty, as the lua processing would turn the
empty array into an empty object, resulting in ElasticSearch errors down
the line.

Signed-off-by: Simon Marechal <bartavelle@gmail.com>