Let multiple limits work at the same time
Fixes #650. The bug was that when a "limit key" can't be built, the
filter decision would be to let pass instead of checking the other
limits.

This means that, for exemple, if a limit counts on a header, and that
header was not present, then subsequent limits would not be checked.

Signed-off-by: Simon Marechal <bartavelle@gmail.com>