Flow control : multiple actions priority
Previously, when several flow control rules would attempt to block a
request, the selected action was undefined. It is now guaranteed that
the action with the highest priority is selected, where the priority is
computed this way:

```rust
    Ban(sub, _) => 10 + sub.atype.priority(),
    Default => 8,
    Challenge => 6,
    Redirect(_) => 4,
    Response(_) => 3,
    RequestHeader(_) => 2,
    Monitor => 1,
```

If several actions have the maximum priority, one of them is selected in
an undefined manner.

Signed-off-by: Simon Marechal <bartavelle@gmail.com>