Fix waf bypass bug
In waf policies, a regex can be set to match values of arguments. When
this value is empty in the UI, it is stored as an empty string in the
configuration file. In that case, the resulting regex would be the empty
regex, matching everything.

This patch fixes that bug, and adds tests.

Signed-off-by: Simon Marechal <bartavelle@gmail.com>