Per-profile hyperscan db
This patch modifies the previous HSDB structure so that it is now per
profile. In benches, with the default configuration, the content filter
accounts for 276µs (approx. 28% of total time). When removing `risk-5`
from the list, that is the majority of signatures, it went down to 127µs
(approx. 16% of total time), which is roughly half the time.

Signed-off-by: Simon Marechal <bartavelle@gmail.com>
Signed-off-by: yitzchake <yitzchake@reblaze.com>